+++ packages/react/src/experimental/Navigation/DaytimePage/index.tsx
@@ -2,6 +2,7 @@ import { F0Button } from "@/components/F0Button"
 import { F0AvatarPerson } from "@/components/avatars/F0AvatarPerson"
 import { F0AvatarPulse } from "@/components/avatars/F0AvatarPulse"
 import { OneSwitch } from "@/experimental/AiChat/OneSwitch"
+import { OneSwitch as OnePromotionSwitch } from "@/experimental/AiPromotionChat/OneSwitch"
 import { useSidebar } from "@/experimental/Navigation/ApplicationFrame/FrameProvider"
 import Menu from "@/icons/app/Menu"
 import { cn } from "@/lib/utils"
@@ -120,6 +121,7 @@ export function DaytimePage({
           </div>
           <div>
             <OneSwitch />
+            <OnePromotionSwitch />
           </div>
         </div>
       )}