+++ packages/react/src/experimental/exports.ts
@@ -2,6 +2,7 @@ import { Component } from "../lib/component/component"
 import { ScrollArea as ScrollAreaComponent } from "./Utilities/ScrollArea"
 
 export * from "./AiChat/exports"
+export * from "./AiPromotionChat/exports"
 export * from "./Banners/exports"
 export * from "./Charts/exports"
 export * from "./Forms/exports"